% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_returns.R
\name{calc_returns}
\alias{calc_returns}
\title{Calculate Portfolio Returns}
\usage{
calc_returns(weighted_universe, universe, portfolio.weight, ret.var,
  points)
}
\arguments{
\item{weighted_universe}{is the universe parameter with characteristic-matching random weights joined}

\item{universe}{is the universe of assets and their characteristics}

\item{portfolio.weight}{is the name of the column which contains the target portfolio weights (for calculating exposures)}

\item{ret.var}{is the name of the column which contains the returns of each asset}

\item{points}{is the number of randomly sampled portfolios}
}
\value{
returns, a data.frame of each portfolio's return
}
\description{
This function calculates the returns of all of the portfolios
considered, adding in a portfolio with even weight across the
universe for comparison
}
\examples{
\dontrun{
returns <- calc_returns(weighted_universe, universe, portfolio.weight, ret.var, points)
}
}
