% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exposures.R
\name{check_exposures}
\alias{check_exposures}
\title{Check Exposures}
\usage{
check_exposures(weighted_universe, match, portfolio.weight, n)
}
\arguments{
\item{weighted_universe}{is the $frame output of portfoliowalkr}

\item{match}{is the vector of column names which contain the
characteristics whose exposures we'd like to compare}

\item{portfolio.weight}{is the name of the column which contains
the target portfolio weights (for calculating exposures)}

\item{n}{is the number of weighted portfolios we'd like to compare}
}
\value{
exposures, a frame of portfolio exposures
}
\description{
This function checks a random number of matched columns
from the universe to demonstrate that exposures are the same
}
\examples{
\dontrun{
exposures <- check_exposures(full_list$frame, match = c('growth', 'size', 'sector'), portfolio.weight = "portfolio", n = 3)
}
}
