% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_small.R
\name{make_small}
\alias{make_small}
\title{Make Small Universe}
\usage{
make_small(universe, match, portfolio.weight)
}
\arguments{
\item{universe}{is the set of stocks available in the universe,
with target weights}

\item{match}{is the vector of column names which contain the
characteristics we'd like to match}

\item{portfolio.weight}{is the name of the column which contains
the target portfolio weights (for calculating exposures)}
}
\value{
list small_set which contains (1) the universe_small with
only relevant assets for matching and (2) a revised match vector
if we no longer want to match on a column
}
\description{
This function creates a simplified universe for cases where
not all factor variables are represented or only numeric values
at the edge of the distribution are represented in the portfolio.
These distributions flatten the polytope and make it impossible for
walkr to properly sample, so they must be removed before sampling
then later reattached.
}
\examples{
\dontrun{
small_set <- make_small(universe, match, portfolio.weight)
}
}
