% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_summary.R
\name{performance_summary}
\alias{performance_summary}
\title{Performance Summary}
\usage{
performance_summary(returns, points)
}
\arguments{
\item{returns}{is a dataframe of portfolio returns}

\item{points}{is the number of randomly sampled portfolios}
}
\value{
summary, a list with figures describing portfolio performance
}
\description{
This function creates a summary table of the target 
portfolio's performance
}
\examples{
\dontrun{
summary <- performance_summary(returns, points)
}
}
