% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_return_distribution.R
\name{plot_return_distribution}
\alias{plot_return_distribution}
\title{Plot Return Distribution}
\usage{
plot_return_distribution(returns, points)
}
\arguments{
\item{returns}{is a dataframe of portfolio returns}

\item{points}{is the number of randomly sampled portfolios}
}
\value{
plot, a histogram of the returns distribution
}
\description{
This function creates a ggplot histogram of the sampled 
portfolio distributions against an evenly weighted portfolio
and the target portfolio
}
\examples{
\dontrun{
plot <- plot_return_distribution(returns, portfolio.weight)
}
}
