% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_a.R
\name{transpose_a}
\alias{transpose_a}
\title{Transpose A}
\usage{
transpose_a(universe_small, match)
}
\arguments{
\item{universe_small}{is the set of stocks available in the 
universe (with redundant assets removed), with target weights}

\item{match}{is the vector of column names which contain the
characteristics we'd like to match}
}
\value{
A, n characteristics in height, for walkr computation
}
\description{
This function transposes the universe parameter, one_hot
encoding any factor variables.
}
\examples{
\dontrun{
A <- transpose_a(universe_small, match)
}
}
