% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose_b.R
\name{transpose_b}
\alias{transpose_b}
\title{Transpose b}
\usage{
transpose_b(universe_small, match, portfolio.weight)
}
\arguments{
\item{universe_small}{is the set of stocks available in the 
universe (with redundant assets removed), with target weights}

\item{match}{is the vector of column names which contain the
characteristics we'd like to match}

\item{portfolio.weight}{is the name of the column which contains
the target portfolio weights (for calculating exposures)}
}
\value{
b, a frame of target portfolio exposures
}
\description{
This function transposes creates parameter b for 
use with walkr. The function calculates the target
portfolio's exposures and returns a transposed frame
for matching
}
\examples{
\dontrun{
b <- transpose_b(universe_small, match, portfolio.weight)
}
}
